+++ demos/editor_entity/Form1.cs
@@ -1153,7 +1153,7 @@ namespace Demos
                 var fileName = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "sample\\gerber\\LED-seven-segment.GBS");
                 if (!File.Exists(fileName)) 
                     return;
-                bool success = EntityFactory.CreateGerber(fileName, SpiralLab.Sirius3.UI.Config.ScannerPenColors[0], out var gerber);
+                bool success = EntityFactory.CreateGerber(fileName, SpiralLab.Sirius3.UI.Config.EntityPenColors[0], out var gerber);
                 Debug.Assert(success);
                 gerber.Translate(rng.NextDouble() * 100.0 - 50, rng.NextDouble() * 100.0 - 50, 0);
                 document.ActivePage?.ActiveLayer?.AddChild(gerber);
@@ -1163,7 +1163,7 @@ namespace Demos
                 var fileName = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "sample\\gerber\\TRF7960_EVM (REV A).TOP");
                 if (!File.Exists(fileName)) 
                     return;
-                bool success = EntityFactory.CreateGerber(fileName, SpiralLab.Sirius3.UI.Config.ScannerPenColors[1], out var gerber);
+                bool success = EntityFactory.CreateGerber(fileName, SpiralLab.Sirius3.UI.Config.EntityPenColors[1], out var gerber);
                 Debug.Assert(success);
                 gerber.Translate(rng.NextDouble() * 100.0 - 50, rng.NextDouble() * 100.0 - 50, 0);
                 document.ActivePage?.ActiveLayer?.AddChild(gerber);